feat(cli): report the provider, the fallback and the failure category
Every command that produces a model now prints the same provider block:
requested and resolved backend, whether an adapter was executed at all and why
not, the host and target architectures as separate facts, the producer and its
artifact digest, the parser family, profile and artifact the registry named, and
the containment the child reported. `info`, the decompile report and both sides
of `diff` build it from one value, so they cannot describe the same run
differently.

Failures now carry a stable category alongside the message. The message is for a
human and may be reworded; `adapter_timeout`, `adapter_malformed_document` and
`registry_no_record` are what a script matches on. `info` stops swallowing an
adapter failure: it prints its report, names the category, and exits nonzero,
because a report with the model fields missing looks exactly like a snapshot
with nothing in it.

`diff` reports each side separately and flags a two-sided run whose sides were
not produced the same way. A heuristic candidate with no name, owner or library
is counted as uncomparable rather than folded into the descriptor `::`, where
every such candidate on both sides collapsed into one entry that read as
"unchanged".

The CLI integration rig moves to a shared `tests/support` so more than one test
binary can drive the packaged prefix.
